requirements.txt
bing_image_downloader==1.1.0
Pillow==8.4.0
imagehash==4.2.1
